+++ frontend/taipy-gui/src/components/Taipy/Field.tsx
@@ -14,6 +14,7 @@
 import React, { lazy, useMemo, Suspense } from "react";
 import Typography from "@mui/material/Typography";
 import Tooltip from "@mui/material/Tooltip";
+import { MathJax } from "better-react-mathjax";
 
 import { formatWSValue } from "../../utils";
 import { getSuffixedClassNames } from "./utils";
@@ -34,23 +35,6 @@ interface TaipyFieldProps extends TaipyBaseProps, TaipyHoverProps {
 const unsetWeightSx = { fontWeight: "unset" };
 
 const Markdown = lazy(() => import("react-markdown"));
-const MathJax = lazy(() => import("better-react-mathjax").then((module) => ({ default: module.MathJax })));
-const MathJaxContext = lazy(() =>
-    import("better-react-mathjax").then((module) => ({ default: module.MathJaxContext }))
-);
-
-const mathJaxConfig = {
-    tex: {
-        inlineMath: [
-            ["$", "$"],
-            ["\\(", "\\)"],
-        ],
-        displayMath: [
-            ["$$", "$$"],
-            ["\\[", "\\]"],
-        ],
-    },
-};
 
 const Field = (props: TaipyFieldProps) => {
     const { id, dataType, format, defaultValue, raw } = props;
@@ -117,19 +101,15 @@ const Field = (props: TaipyFieldProps) => {
                         {value}
                     </span>
                 ) : mode == "latex" ? (
-                    <Suspense fallback={<div>Loading LaTex...</div>}>
-                        <MathJaxContext config={mathJaxConfig}>
-                            <MathJax
-                                className={`${className} ${getSuffixedClassNames(
-                                    className,
-                                    "-latex"
-                                )} ${getComponentClassName(props.children)}`}
-                                id={id}
-                            >
-                                {value}
-                            </MathJax>
-                        </MathJaxContext>
-                    </Suspense>
+                    <MathJax
+                        className={`${className} ${getSuffixedClassNames(className, "-latex")} ${getComponentClassName(
+                            props.children
+                        )}`}
+                        id={id}
+                        dynamic={true}
+                    >
+                        {value}
+                    </MathJax>
                 ) : (
                     <Typography
                         className={`${className} ${
